ICU-22917 Separate methods in three classes
